test: split sequence tests into test_parser_bytecode and test_parser_vm
Per review feedback:
- Remove test_sequence.cpp; move SequenceBytecodeTest into
  test_parser_bytecode.cpp and SequenceVMTest into test_parser_vm.cpp
- Eliminate the duplicated trim() helper (already defined in
  test_parser_bytecode.cpp)
- Add comment explaining that Chunk::addConstant does an O(N) dedup scan
  so the repeated 2.0 constant shares slot 0 in InExpr_ListMembership
- Fix clang-format line-length violations in comments
