+++ test/CMakeLists.txt
@@ -37,7 +37,6 @@ add_executable(test_gc_regression test_gc_regression.cpp test_harness.cpp ${FULL
 add_executable(test_inheritance test_inheritance.cpp test_harness.cpp ${FULL_SRCS})
 add_executable(test_list test_list.cpp test_harness.cpp ${FULL_SRCS})
 add_executable(test_list_gc test_list.cpp test_harness.cpp ${FULL_SRCS})
-add_executable(test_sequence test_sequence.cpp test_harness.cpp ${FULL_SRCS})
 
 target_include_directories(test_main PRIVATE ${PROJECT_SOURCE_DIR}/src)
 target_include_directories(test_scanner PRIVATE ${PROJECT_SOURCE_DIR}/src)
@@ -54,7 +53,6 @@ target_include_directories(test_gc_regression PRIVATE ${PROJECT_SOURCE_DIR}/src)
 target_include_directories(test_inheritance PRIVATE ${PROJECT_SOURCE_DIR}/src)
 target_include_directories(test_list PRIVATE ${PROJECT_SOURCE_DIR}/src)
 target_include_directories(test_list_gc PRIVATE ${PROJECT_SOURCE_DIR}/src)
-target_include_directories(test_sequence PRIVATE ${PROJECT_SOURCE_DIR}/src)
 
 target_compile_definitions(test_stress_gc PRIVATE LOXPP_STRESS_GC)
 target_compile_definitions(test_gc_regression PRIVATE LOXPP_STRESS_GC)
@@ -74,7 +72,6 @@ target_link_libraries(test_gc_regression PRIVATE GTest::GTest GTest::Main)
 target_link_libraries(test_inheritance PRIVATE GTest::GTest GTest::Main)
 target_link_libraries(test_list PRIVATE GTest::GTest GTest::Main)
 target_link_libraries(test_list_gc PRIVATE GTest::GTest GTest::Main)
-target_link_libraries(test_sequence PRIVATE GTest::GTest GTest::Main)
 
 add_test(NAME TestMain COMMAND test_main)
 add_test(NAME TestScanner COMMAND test_scanner)
@@ -91,4 +88,3 @@ add_test(NAME TestGcRegression COMMAND test_gc_regression)
 add_test(NAME TestInheritance COMMAND test_inheritance)
 add_test(NAME TestList COMMAND test_list)
 add_test(NAME TestListGC COMMAND test_list_gc)
-add_test(NAME TestSequence COMMAND test_sequence)
